requirements.txt
opencv-python-headless==4.5.4.58
numpy
Pillow
Flask==1.1.1